Extract async job pool into a separate class
This is the first step towards 1 runner + N pools model (vs current N runners model)
